Add Environmental Message Type8 Dac367 Fid33
This implements the ability to generate Environmental Messages as
documented at
https://www.e-navigation.nl/sites/default/files/asm_files/em_version_release_3-23mar15_0.pdf
This makes a few notable changes in accomplishing this. A set of
SensorReport classes are added which currently inherit from the Payload
class. This brings its own difficulties, but for now while I am testing
I think it is sufficient.
I have also added two new fields to bit width: scale and offset. Scale
is meant to be used to define the scale of units, and offset is meant
to offset values that do not start counting at zero. For example, in
sensor report 8, salinity, a value of 0 for the temperature corresponds
to -10 celcius.